backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
jinja2==3.1.4
python-dotenv==1.0.1
pydantic==2.9.2
sqlmodel==0.0.21
paramiko==3.4.0
httpx==0.27.2
